+++ components/side-drawer/test/side-drawer.test.js
@@ -11,7 +11,7 @@ chai.use(chaiDomDiff);
 
 const COMPONENT_NAME = 'vwc-side-drawer';
 
-function animateModal(drawerElement) {
+function animateDrawer(drawerElement) {
 	const event = new Event('transitionend');
 	drawerElement.dispatchEvent(event);
 }
@@ -53,15 +53,12 @@ describe('Side-drawer', () => {
 			expect(actualElement.hasTopBar, 'hasTopBar should be undefined')
 				.to
 				.equal(undefined);
-			expect(actualElement.absolute, 'absolute should be false')
-				.to
-				.equal(false);
 		});
 	});
 
 	describe('Side drawer attributes', () => {
 		it('should reflect from attribute to property', async () => {
-			const COMPONENT_PROPERTIES = ['open', 'alternate', 'hasTopBar', 'absolute'];
+			const COMPONENT_PROPERTIES = ['open', 'alternate', 'hasTopBar'];
 			for await (const property of COMPONENT_PROPERTIES) {
 				const [actualElement] = addElement(
 					textToDomToParent(`<${COMPONENT_NAME} ${property}></${COMPONENT_NAME}>`)
@@ -95,7 +92,7 @@ describe('Side-drawer', () => {
 				textToDomToParent(`<${COMPONENT_NAME} type="modal"></${COMPONENT_NAME}>`)
 			);
 		});
-		it('should fire opened event after animation completes and open is true', async () => {
+		it('should fire opened and trapFocus events after animation completes and open is true', async () => {
 			const onOpened = chai.spy();
 			const onFocusTrapped = chai.spy();
 
@@ -113,14 +110,14 @@ describe('Side-drawer', () => {
 			});
 
 			sideDrawerEl.open = true;
-			animateModal(sideDrawerEl);
+			animateDrawer(sideDrawerEl);
 
 			await eventListenerPromise;
 			onOpened.should.have.been.called();
 			onFocusTrapped.should.have.been.called();
 		});
 
-		it('should fire closed event after animation completes and open is false', async () => {
+		it('should fire closed and releaseFocus events after animation completes and open is false', async () => {
 			const onClosed = chai.spy();
 			const onFocusReleased = chai.spy();
 
@@ -139,7 +136,7 @@ describe('Side-drawer', () => {
 			});
 
 			sideDrawerEl.open = false;
-			animateModal(sideDrawerEl);
+			animateDrawer(sideDrawerEl);
 
 			await eventListenerPromise;
 			onClosed.should.have.been.called();
@@ -151,8 +148,7 @@ describe('Side-drawer', () => {
 
 			sideDrawerEl.open = true;
 			await sideDrawerEl.updateComplete;
-			const scrim = sideDrawerEl.shadowRoot.querySelector('.vvd-side-drawer--scrim');
-
+			const scrim = sideDrawerEl.shadowRoot.querySelector('.side-drawer--scrim');
 
 			const eventListenerPromise = new Promise((res) => {
 				sideDrawerEl.addEventListener('closed', () => {
@@ -162,18 +158,17 @@ describe('Side-drawer', () => {
 			});
 
 			scrim?.click();
-			animateModal(sideDrawerEl);
+			animateDrawer(sideDrawerEl);
 
 			await eventListenerPromise;
 			onClosed.should.have.been.called();
 		});
 
-		it('should fire closed event after pressing escape on the drawer', async () => {
+		it('should fire closed event after pressing escape on the document', async () => {
 			const onClosed = chai.spy();
 
 			sideDrawerEl.open = true;
 			await sideDrawerEl.updateComplete;
-
 
 			const eventListenerPromise = new Promise((res) => {
 				sideDrawerEl.addEventListener('closed', () => {
@@ -183,9 +178,57 @@ describe('Side-drawer', () => {
 			});
 
 			const keyboardEvent = new KeyboardEvent('keydown', { key: 'Escape' });
-			sideDrawerEl.dispatchEvent(keyboardEvent);
+			document.dispatchEvent(keyboardEvent);
 
-			animateModal(sideDrawerEl);
+			animateDrawer(sideDrawerEl);
+
+			await eventListenerPromise;
+			onClosed.should.have.been.called();
+		});
+	});
+
+	describe('Dismissible side drawer events', () => {
+		let sideDrawerEl;
+
+		beforeEach(function () {
+			[sideDrawerEl] = addElement(
+				textToDomToParent(`<${COMPONENT_NAME} type="dismissible"></${COMPONENT_NAME}>`)
+			);
+		});
+		it('should fire opened event after animation completes and open is true', async () => {
+			const onOpened = chai.spy();
+
+			await sideDrawerEl.updateComplete;
+
+			const eventListenerPromise = new Promise((res) => {
+				sideDrawerEl.addEventListener('opened', () => {
+					onOpened();
+					res();
+				});
+			});
+
+			sideDrawerEl.open = true;
+			animateDrawer(sideDrawerEl);
+
+			await eventListenerPromise;
+			onOpened.should.have.been.called();
+		});
+
+		it('should fire closed event after animation completes and open is false', async () => {
+			const onClosed = chai.spy();
+
+			sideDrawerEl.open = true;
+			await sideDrawerEl.updateComplete;
+
+			const eventListenerPromise = new Promise((res) => {
+				sideDrawerEl.addEventListener('closed', () => {
+					onClosed();
+					res();
+				});
+			});
+
+			sideDrawerEl.open = false;
+			animateDrawer(sideDrawerEl);
 
 			await eventListenerPromise;
 			onClosed.should.have.been.called();
@@ -197,7 +240,6 @@ describe('Side-drawer', () => {
 			const [actualElement] = addElement(
 				textToDomToParent(`<${COMPONENT_NAME}></${COMPONENT_NAME}>`)
 			);
-			actualElement.open = false;
 			actualElement.show();
 			expect(actualElement.open)
 				.to
@@ -208,9 +250,8 @@ describe('Side-drawer', () => {
 	describe(`hide`, function () {
 		it(`should set "open" to false`, function () {
 			const [actualElement] = addElement(
-				textToDomToParent(`<${COMPONENT_NAME}></${COMPONENT_NAME}>`)
+				textToDomToParent(`<${COMPONENT_NAME} open></${COMPONENT_NAME}>`)
 			);
-			actualElement.open = true;
 			actualElement.hide();
 			expect(actualElement.open)
 				.to