+++ CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.18.1](https://github.com/Vonage/vivid/compare/v2.18.0...v2.18.1) (2021-08-25)
+
+
+### Features
+
+* **vwc-side-drawer:** remove absolute support ([#1029](https://github.com/Vonage/vivid/issues/1029)) ([b00ded7](https://github.com/Vonage/vivid/commit/b00ded72111a63958d4370609943c160782cf1a2))
+
+
+
+
+
+# [2.18.0](https://github.com/Vonage/vivid/compare/v2.17.0...v2.18.0) (2021-08-24)
+
+
+### Bug Fixes
+
+* **side-drawer:** refactor side drawer ([#1022](https://github.com/Vonage/vivid/issues/1022)) ([2788c2b](https://github.com/Vonage/vivid/commit/2788c2bf91e2d753f515d950fb09bfcd0c68657d))
+* **textarea:** textarea scroll now working ([#1030](https://github.com/Vonage/vivid/issues/1030)) ([86ae56e](https://github.com/Vonage/vivid/commit/86ae56ed40a3548e9e7ea88f11830f602baa7e42))
+* **vwc-side-drawer:** handle scrim a11y violation ([#1024](https://github.com/Vonage/vivid/issues/1024)) ([518496e](https://github.com/Vonage/vivid/commit/518496e831d5ec3d16299c7a3bcfb3f8d9956a46))
+
+
+### Features
+
+* **card:** vwc card first draft ([#1001](https://github.com/Vonage/vivid/issues/1001)) ([1d41627](https://github.com/Vonage/vivid/commit/1d41627aea4dca84da6e39b73a435a13d3664769))
+* **textfield:** add the no-actions-sync attribute ([#1013](https://github.com/Vonage/vivid/issues/1013)) ([850297f](https://github.com/Vonage/vivid/commit/850297ffb224b1badf6b17b144c7b75e7bffcf61))
+* **vwc-side-drawer:** add dismissible support ([#1025](https://github.com/Vonage/vivid/issues/1025)) ([65c38d1](https://github.com/Vonage/vivid/commit/65c38d130dfdf8b9d2b1c985c498a8f3c7143627))
+* **vwc-side-drawer:** add modal support ([#1002](https://github.com/Vonage/vivid/issues/1002)) ([02e2782](https://github.com/Vonage/vivid/commit/02e278210a9d5f9a612dd8b7f9ae5f462a0fd477))
+
+
+
+
+
 # [2.17.0](https://github.com/Vonage/vivid/compare/v2.16.1...v2.17.0) (2021-08-09)